refactor: Cleanup code and add comments.
