+++ lib/coverage-map.js
@@ -1,0 +1,22 @@
+// Coverage Map
+// ============
+
+var coverageMap = {}
+
+function add (coverageObj) {
+  coverageMap[coverageObj.path] = coverageObj
+}
+
+function get () {
+  return coverageMap
+}
+
+function reset () {
+  coverageMap = {}
+}
+
+module.exports = {
+  add: add,
+  get: get,
+  reset: reset
+}
