+++ lib/preprocessor.js
@@ -8,9 +8,9 @@
 
 var istanbul = require('istanbul')
 var minimatch = require('minimatch')
-var globalSourceCache = require('./sourceCache')
+var globalSourceCache = require('./source-cache')
 var extend = require('util')._extend
-var coverageMap = require('./coverageMap')
+var coverageMap = require('./coverage-map')
 
 // Regexes
 // -------
@@ -35,7 +35,7 @@ function createCoveragePreprocessor (logger, helper, basePath, reporters, covera
     includeAllSources = coverageReporter.includeAllSources === true
   }
 
-  var sourceCache = globalSourceCache.getByBasePath(basePath)
+  var sourceCache = globalSourceCache.get(basePath)
 
   var instrumentersOptions = _.reduce(instrumenters, function getInstumenterOptions (memo, instrument, name) {
       memo[name] = {}
