+++ lib/reporter.js
@@ -9,8 +9,8 @@
 var path = require('path')
 var istanbul = require('istanbul')
 
-var globalSourceCache = require('./sourceCache')
-var coverageMap = require('./coverageMap')
+var globalSourceCache = require('./source-cache')
+var coverageMap = require('./coverage-map')
 var SourceCacheStore = require('./source-cache-store')
 
 // TODO(vojta): inject only what required (config.basePath, config.coverageReporter)
@@ -29,7 +29,7 @@ var CoverageReporter = function (rootConfig, helper, logger) {
   var config = rootConfig.coverageReporter || {}
   var basePath = rootConfig.basePath
   var reporters = config.reporters
-  var sourceCache = globalSourceCache.getByBasePath(basePath)
+  var sourceCache = globalSourceCache.get(basePath)
   var includeAllSources = config.includeAllSources === true
 
   if (config.watermarks) {
