--- lib/sourceCache.js
@@ -1,5 +1,0 @@
-var cacheByBasePath = {}
-
-exports.getByBasePath = function (basePath) {
-  return cacheByBasePath[basePath] ? cacheByBasePath[basePath] : (cacheByBasePath[basePath] = {})
-}
